requirements.txt
mysql-connector-python==8.2.0
sqlalchemy==2.0.23
pandas==2.1.4
python-dotenv==1.0.0
numpy==1.24.3
tqdm==4.66.1
pymysql==1.1.0
streamlit
